feat: adds group by (bg or text) feature
resolves #31
