+++ src/components/ResultsForm/index.js
@@ -3,28 +3,8 @@ import { Subscribe } from 'unstated';
 import { Select, Checkbox } from 'cauldron-react';
 import PaletteContainer from '../../containers/PaletteContainer';
 import TextInput from '../TextInput';
+import { fontSize, fontWeight, groupBy } from './options';
 import './index.css';
-const fontSizeOptions = [
-  { label: '10pt', value: 10 },
-  { label: '11pt', value: 11 },
-  { label: '12pt', value: 12 },
-  { label: '13pt', value: 13 },
-  { label: '14pt', value: 14 },
-  { label: '15pt', value: 15 },
-  { label: '16pt', value: 16 },
-  { label: '17pt', value: 17 },
-  { label: '18pt', value: 18 },
-  { label: '19pt', value: 19 },
-  { label: '20pt', value: 20 }
-];
-const fontWeightOptions = [
-  { value: 'normal', label: 'Normal' },
-  { value: 'bold', label: 'Bold' }
-];
-const groupByOptions = [
-  { value: 'background', label: 'Background color' },
-  { value: 'text', label: 'Text color' }
-];
 
 export default class ResultsForm extends Component {
   render() {
@@ -51,7 +31,7 @@ export default class ResultsForm extends Component {
                 selectedId="selected-font-size-option"
                 value={15}
                 className="font-size-select"
-                options={fontSizeOptions}
+                options={fontSize}
                 onSelect={({ value }) => {
                   updateResultsSettings({ fontSize: value });
                 }}
@@ -62,7 +42,7 @@ export default class ResultsForm extends Component {
                 selectedId="selected-font-weight-option"
                 value="normal"
                 className="font-weight-select"
-                options={fontWeightOptions}
+                options={fontWeight}
                 onSelect={({ value }) => {
                   updateResultsSettings({ fontWeight: value });
                 }}
@@ -73,7 +53,7 @@ export default class ResultsForm extends Component {
                 selectedId="selected-group-by-option"
                 value="background"
                 className="group-by-select"
-                options={groupByOptions}
+                options={groupBy}
                 onSelect={({ value }) => {
                   updateResultsSettings({ groupBy: value });
                 }}
