+++ src/components/ResultsForm/options.js
@@ -1,0 +1,21 @@
+export const fontSize = [
+  { label: '10pt', value: 10 },
+  { label: '11pt', value: 11 },
+  { label: '12pt', value: 12 },
+  { label: '13pt', value: 13 },
+  { label: '14pt', value: 14 },
+  { label: '15pt', value: 15 },
+  { label: '16pt', value: 16 },
+  { label: '17pt', value: 17 },
+  { label: '18pt', value: 18 },
+  { label: '19pt', value: 19 },
+  { label: '20pt', value: 20 }
+];
+export const fontWeight = [
+  { value: 'normal', label: 'Normal' },
+  { value: 'bold', label: 'Bold' }
+];
+export const groupBy = [
+  { value: 'background', label: 'Background color' },
+  { value: 'text', label: 'Text color' }
+];
